Remove residuals from previous sem-type integration logics
