--- .github/workflows/daily_build_semtypes.yml
@@ -1,20 +1,0 @@
-name: Daily unit-test run with SemTypes activated
-
-on:
-  workflow_dispatch:
-  schedule:
-    - cron: "30 11 * * *"
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    timeout-minutes: 30 # random number
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up JDK 11
-        uses: actions/setup-java@v3
-        with:
-          java-version: '11'
-          distribution: 'temurin'
-      - name: Build with Gradle
-        run: ./gradlew semtypesUnitTest configure-on-demand --no-daemon
