+++ compiler/ballerina-lang/src/main/java/io/ballerina/compiler/api/impl/type/builders/BallerinaSingletonTypeBuilder.java
@@ -23,7 +23,7 @@ import io.ballerina.compiler.api.impl.symbols.AbstractTypeSymbol;
 import io.ballerina.compiler.api.impl.symbols.TypesFactory;
 import io.ballerina.compiler.api.symbols.SingletonTypeSymbol;
 import io.ballerina.compiler.api.symbols.TypeSymbol;
-import org.wso2.ballerinalang.compiler.semantics.analyzer.SemTypeResolver;
+import org.wso2.ballerinalang.compiler.semantics.analyzer.SemTypeHelper;
 import org.wso2.ballerinalang.compiler.semantics.model.SymbolTable;
 import org.wso2.ballerinalang.compiler.semantics.model.symbols.BTypeSymbol;
 import org.wso2.ballerinalang.compiler.semantics.model.symbols.SymTag;
@@ -77,7 +77,7 @@ public class BallerinaSingletonTypeBuilder implements TypeBuilder.SINGLETON {
                 symTable.builtinPos, COMPILED_SOURCE);
 
         BFiniteType finiteType = BFiniteType.newSingletonBFiniteType(finiteTypeSymbol,
-                SemTypeResolver.resolveSingletonType(valueLiteral));
+                SemTypeHelper.resolveSingletonType(valueLiteral));
         finiteTypeSymbol.type = finiteType;
         SingletonTypeSymbol singletonTypeSymbol = (SingletonTypeSymbol) typesFactory.getTypeDescriptor(finiteType,
                 finiteTypeSymbol, true);
