+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/semantics/analyzer/ConditionResolver.java
@@ -52,10 +52,10 @@ public class ConditionResolver {
                     return value == Boolean.TRUE ? symTable.trueType : symTable.falseType;
                 }
                 return BFiniteType.newSingletonBFiniteType(null,
-                        SemTypeResolver.resolveSingletonType((BLangLiteral) condition));
+                        SemTypeHelper.resolveSingletonType((BLangLiteral) condition));
             case NUMERIC_LITERAL:
                 return BFiniteType.newSingletonBFiniteType(null,
-                        SemTypeResolver.resolveSingletonType((BLangLiteral) condition));
+                        SemTypeHelper.resolveSingletonType((BLangLiteral) condition));
             case TYPE_TEST_EXPR:
                 BLangTypeTestExpr typeTestExpr = (BLangTypeTestExpr) condition;
                 BType exprType = typeTestExpr.expr.getBType();
