+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/semantics/model/types/BType.java
@@ -22,7 +22,7 @@ import io.ballerina.types.SemType;
 import org.ballerinalang.model.Name;
 import org.ballerinalang.model.types.TypeKind;
 import org.ballerinalang.model.types.ValueType;
-import org.wso2.ballerinalang.compiler.semantics.analyzer.SemTypeResolver;
+import org.wso2.ballerinalang.compiler.semantics.analyzer.SemTypeHelper;
 import org.wso2.ballerinalang.compiler.semantics.model.TypeVisitor;
 import org.wso2.ballerinalang.compiler.semantics.model.symbols.BTypeSymbol;
 import org.wso2.ballerinalang.compiler.util.Names;
@@ -109,7 +109,7 @@ public class BType implements ValueType {
     }
 
     public boolean isNullable() {
-        return Core.containsNil(SemTypeResolver.semTypeComponent(this));
+        return Core.containsNil(SemTypeHelper.semTypeComponent(this));
     }
 
     public <T, R> R accept(BTypeVisitor<T, R> visitor, T t) {
